requirements.txt
chromadb>=0.4.0
pyyaml>=6.0
requests>=2.31.0
